+++ CHANGELOG.md
@@ -1,6 +1,10 @@
 # Yall changelog
 
 ## 1.0.0.pre.2 
+- dot folders included by default
+- `clean-up` option
+- handle concurrent run cache error https://github.com/yarnpkg/yarn/issues/683
+
 
 ## 1.0.0.pre.1 (2017-05-02) 
 - fix `--folders`